requirements.txt
asgiref==3.5.2
Django==3.2.16
djangorestframework==3.14.0
django-cors-headers==3.13.0
django_filter==22.1
psycopg2-binary==2.9.5
gunicorn==20.1.0
djoser==2.1.0
python-dotenv
pytest==7.2.0
pytest-cov
pytest-django
pytest-mock